+++ whatsnew.md
@@ -1,3 +1,4 @@
 - ERAM
   - Added support for point outs (QP commands)
   - Fixed bug with placement of dwell boxes with leader-line length 0
+  - Fixed bugs in LA/LB commands, added (missing) LC command